That was actually quite easy.
Make a program where the user is asked about how many numbers he/she
wants to input and have processed.
Create an array with the size of what the user answered.
Ask the user to input numbers, one number at the time, and put it into
the array.

Find the following in the array:
• The highest number
• The lowest number
• The sum of the numbers
• The average of the numbers
• The total amout of numbers in the array.
Print these...